feat: emails when admin removes volunteer or delete checkin (#230)
* add admin remove volunteer from checkin email

* add volunteer email when checkin deleted by admin

* send volunteer email when admin deletes checkin by range

* change email link

* use CancellationEmail util func

* Edited email

+++ backend/typescript/services/implementations/checkInService.ts
@@ -21,6 +21,7 @@ import IVolunteerService from "../interfaces/volunteerService";
 import VolunteerService from "./volunteerService";
 import ContentService from "./contentService";
 import {
+  cancellationEmail,
   emailFooter,
   emailHeader,
   formatCheckinShiftInformation,
@@ -335,6 +336,42 @@ class CheckInService implements ICheckInService {
     }
   }
 
+  async sendVolunteerAdminCancelCheckInEmail(
+    volunteerId: string,
+    checkIn: CheckIn,
+  ): Promise<void> {
+    if (!this.emailService) {
+      const errorMessage =
+        "Attempted to call sendVolunteerAdminCancelCheckInEmail but this instance of CheckInService does not have an EmailService instance";
+      Logger.error(errorMessage);
+      throw new Error(errorMessage);
+    }
+    try {
+      const volunteerService: IVolunteerService = new VolunteerService();
+      const { firstName, email } = await volunteerService.getVolunteerById(
+        volunteerId,
+      );
+      const startDayString: string = dayjs
+        .tz(checkIn.start_date)
+        .format("dddd, MMMM D");
+      const startTimeString: string = dayjs
+        .tz(checkIn.start_date)
+        .format("h:mm A");
+      const endTimeString: string = dayjs.tz(checkIn.end_date).format("h:mm A");
+      const volunteerMainLine = `Your scheduled shift for ${startDayString} from ${startTimeString} to ${endTimeString} has been cancelled.`;
+      this.emailService.sendEmail(
+        email,
+        `Cancellation Notice: Fridge Check-in for ${startDayString} at ${startTimeString}`,
+        cancellationEmail(volunteerMainLine, firstName, true),
+      );
+    } catch (error) {
+      Logger.error(
+        `Failed to generate email to confirm volunteer cancellation by admin for check-in shift for volunteer with id ${volunteerId}`,
+      );
+      throw error;
+    }
+  }
+
   async updateCheckInById(
     checkInId: string,
     checkIn: UpdateCheckInDTO,
@@ -365,33 +402,50 @@ class CheckInService implements ICheckInService {
         notes: updatedCheckIn.notes,
         isAdmin: updatedCheckIn.is_admin,
       };
-
       // send volunteer confirmation email
       if (
         Object.prototype.hasOwnProperty.call(checkIn, "volunteerId") &&
-        updatedCheckIn.volunteer_id !== null
+        updatedCheckIn.volunteer_id
       ) {
+        // Volunteer signing up case
         this.sendVolunteerCheckInSignUpConfirmationEmail(
-          checkIn.volunteerId!,
+          String(updatedCheckIn.volunteer_id),
           updatedCheckInDTO,
           false,
         );
         this.sendVolunteerCheckInSignUpConfirmationEmail(
-          checkIn.volunteerId!,
+          String(updatedCheckIn.volunteer_id),
           updatedCheckInDTO,
           true,
         );
       } else if (
+        Object.prototype.hasOwnProperty.call(checkIn, "isAdmin") &&
         Object.prototype.hasOwnProperty.call(checkIn, "volunteerId") &&
-        updatedCheckIn.volunteer_id === null
+        !updatedCheckIn.volunteer_id &&
+        !updatedCheckIn.is_admin &&
+        oldCheckIn &&
+        oldCheckIn.volunteer_id
       ) {
+        // Admin removing volunteer case
+        this.sendVolunteerAdminCancelCheckInEmail(
+          String(oldCheckIn.volunteer_id),
+          oldCheckIn,
+        );
+      } else if (
+        Object.prototype.hasOwnProperty.call(checkIn, "volunteerId") &&
+        !updatedCheckIn.volunteer_id &&
+        !updatedCheckIn.is_admin &&
+        oldCheckIn &&
+        oldCheckIn.volunteer_id
+      ) {
+        // Volunteer removing themselves
         this.sendVolunteerCancelCheckInEmail(
-          String(oldCheckIn?.volunteer_id),
+          String(oldCheckIn.volunteer_id),
           updatedCheckInDTO,
           false,
         );
         this.sendVolunteerCancelCheckInEmail(
-          String(oldCheckIn?.volunteer_id),
+          String(oldCheckIn.volunteer_id),
           updatedCheckInDTO,
           true,
         );
@@ -407,11 +461,21 @@ class CheckInService implements ICheckInService {
 
   async deleteCheckInById(id: string): Promise<void> {
     try {
+      const oldCheckIn = await CheckIn.findOne({ where: { id } });
+      if (!oldCheckIn) {
+        throw new Error(`checkIn with id ${id} not found.`);
+      }
       const numDestroyed = await CheckIn.destroy({
         where: { id: Number(id) },
       });
       if (numDestroyed <= 0) {
         throw new Error(`checkin with id ${id} was not deleted.`);
+      }
+      if (oldCheckIn.volunteer_id) {
+        this.sendVolunteerAdminCancelCheckInEmail(
+          String(oldCheckIn.volunteer_id),
+          oldCheckIn,
+        );
       }
     } catch (error) {
       Logger.error(
@@ -429,6 +493,16 @@ class CheckInService implements ICheckInService {
     const endDateRange = new Date(endDate);
 
     try {
+      const oldDestroyedCheckIns = await CheckIn.findAll({
+        where: {
+          start_date: {
+            [Op.gte]: startDateRange,
+          },
+          end_date: {
+            [Op.lte]: endDateRange,
+          },
+        },
+      });
       const numsDestroyed = await CheckIn.destroy({
         where: {
           start_date: {
@@ -444,6 +518,14 @@ class CheckInService implements ICheckInService {
           `checkins between start date ${startDate} and end date ${endDate} were not deleted.`,
         );
       }
+      oldDestroyedCheckIns.forEach((oldCheckIn) => {
+        if (oldCheckIn.volunteer_id) {
+          this.sendVolunteerAdminCancelCheckInEmail(
+            String(oldCheckIn.volunteer_id),
+            oldCheckIn,
+          );
+        }
+      });
     } catch (error) {
       Logger.error(
         `Failed to delete checkins by start and end date range. Reason = ${getErrorMessage(
